implement amm deposits/withdrawals
In order to deposit/withdraw, user first needs to query
the api to find out current ration for given pool. This is done
by sending `GET /v1/trading/amm/fundingOrderData` with `pool, `token`
and `amount`. If the `token` if one of the pool token, the response
will contain AmmFundingOrder data corresponding to a deposit
of the given amount of token specified, and a required amount of
the other pool token. The result of `GET fundingOrderData` call
can be inspected or presented to the user for approval and then
can be passed to `POST /v1/trading/amm/fundingOrder` for execution.
Withdrawals are done the same way, however in the call to `GET fundingOrderData`
the token needs to be set to the LP token for given pool, and the amount
to the amount of LP token the user is going to return to the pool.
The response of `GET fundingOrderData` will contain information about
how much of each pool token the user will get in return.

NOTE: amm deposits/withdrawals are performed on L2 by executing 2
starkware limit orders, one for each pool token. The LP token received
(or given away) by the user are divided equally between these 2 orders.
The orders need to be signed with users stark private key (which is
done automatically by `src/api/postAmmFundingOrder.js`, using
`dvfStarkProvider`, hence no additional method is needed to sign using
ledger).